Finish get purchase order list feature

+++ src/main/java/com/group3/apiserver/dto/purchase/PurchaseOrderDTO.java
@@ -4,7 +4,8 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
 
-import java.sql.Date;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @Data
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
@@ -12,5 +13,5 @@ public class PurchaseOrderDTO {
     private Integer PoNo;
     private Date purchaseDate;
     private String status;
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
 }
